Prefix set optimizer
This patch optimizes prefix sets like
[ 10.0.0.0/24, 10.0.1.0/24, 10.0.2.0/24, 10.0.3.0/24 ]
into
[ 10.0.0.0/22{24,24} ]

This should improve config loading speed with large config files.

Use only in config-check mode; the appropriate flag is -O.
Beware. The output is written directly on stdout.

This patch is not intended to be ever merged with mainline in this form.
Instead, we should fix bad lexer performance and also merge the
merge-able prefixes on-the-fly during trie creation.