Nest: Changed struct network * in struct rte to const net_addr *. [WIP]
This allows filtering the route without even touching the table on
import. Therefore rejected routes don't mess with the table in any way.
Only accepted routes are linked to the table.

Kernel protocol still uses a dirty hack to get routes synchronized. This
should be sanitized in future to allow proper concurrency.