Conversion of events to coroutines
Instead of running the events in the main thread, we run them in their
separate threads as coroutines.

For now, there is still the_bird_lock which needs to be split to allow
concurrent thread execution.

Running an event directly is deprecated (there is one warning in BGP)
and should be removed in near future.

Re-scheduling an event from itself is deprecated as well. Instead, the
event should get to an appropriate cancellation point and call
ev_suspend() there. Conversion to this policy will be also made in near
future.